マインスイーパーをclass化
