+++ lib/temtanbot.js
@@ -10,6 +10,10 @@ module.exports = class Temtan{
       return;
     }
     this.openWeatherMap = new OpenWeatherMap(openWeatherMapToken);
+
+    // Minesweepe
+    const Mine = require("./mine.js");
+    this.mine = new Mine();
     // 雑多なフラグ処理
     this.jankenFlag = false;
     this.ohanashiFlag = false;
@@ -140,5 +144,26 @@ module.exports = class Temtan{
         reject(msg);
       });
     });
+  }
+  play_mine(message_content){
+    let mine_cmd = message_content.split(" ");
+    if(mine_cmd[1] === "show"){
+      return "```" + JSON.stringify(this.mine) + "```";
+    }
+    else if(mine_cmd[1] === "reset"){
+      let set_num = Number(mine_cmd[2]);
+      let msg = "強制リセットをかけます！\n";
+      msg += this.mine.start_game(set_num);
+      return msg;
+    }
+    else if(!this.mine.playing){
+      let set_num = Number(mine_cmd[1]);
+      return this.mine.start_game(set_num);
+    }
+    else{
+      let now_msg = this.mine.open_board(mine_cmd[1], mine_cmd[2]);
+      let board_status = this.mine.show_board_discord();
+      return board_status + "\n" + now_msg;
+    }
   }
 }
